feat: add tests for the npx command
